requirements.txt
numpy
faiss-cpu
sentence-transformers
ollama
openai
gradio
PyPDF2